requirements.txt
# Core dependencies
aiohttp>=3.13.2
anthropic>=0.72.1
openai>=2.8.0
requests>=2.32.5
zai-sdk>=0.0.4,<0.1.0

# Web and data processing
lxml>=6.0.2
beautifulsoup4>=4.12.2
html5lib>=1.1

# Development and testing
pytest>=8.4.2
pytest-asyncio>=1.2.0
pytest-cov>=7.0.0
coverage>=7.11.1

# Configuration and environment
python-dotenv>=1.1.1
pydantic>=2.0.0,<2.12.0
pyyaml>=6.0.3

# File processing
python-docx>=1.2.0
python-pptx>=1.0.2
pypdf>=4.3.1
openpyxl>=3.1.5
xlsxwriter>=3.2.9

# Visualization and design
matplotlib>=3.10.7
pillow>=11.0.0,<12.0.0

# Communication protocols
websockets>=11.0.0,<15.0.0
mcp>=1.19.0

# Utilities
tqdm>=4.67.1
click>=8.3.0
colorama>=0.4.6